pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "lyrics-overlay"
version = "0.1.0"
description = "macOS desktop lyrics sync overlay"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "pyobjc-core>=9.0",
    "pyobjc-framework-Cocoa>=9.0",
    "pygame>=2.5",
    "requests>=2.31",
    "mutagen>=1.47",
]

[project.scripts]
lyrics-overlay = "lyrics_overlay.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["lyrics_overlay*"]
